Abandoned attempt to spread contact points.
Does a weighted shuffle of the triangles in a mesh when the mesh is
loaded in an attempt to spread contact points evenly over the mesh
surface. Unfortunately,this is not effective because the collision
algorithm does not process the triangles in order. It puts them in a
spatial tree for performance reasons and this results in
nearby triangles being near each other in the evaluation order.